playground: moved apps-list usage to app-home
